Building out UI for Pokedex
The code previously used to visually make edits to Pokemon icons has been moved to the PokemonEditor script. I'm leaving that alone for the time being, and working on a copy of that code in the PokemonUI file. I now have Pokemon stats displaying as progress bar objects within a frame, and now need to finish implementing a dropdown menu to choose which game series to view. Once that is in place, I'll need to do some more database work, in order to have each game only showing Pokemon and stats from that game/generation. I think I might need to rework the stats in the database as well, but that'll come later.