ERD work and database adjustments
Made a basic ERD for the SQLite database, and adjusted the database to better handle current and future data that need to be added. Also cleaned up a bit of the UI code. The data model is really good for seeing complete regional pokedex data for each game, but is unfortunately not as good for viewing each game's national dex. Not sure yet how I'll accomplish this, but I need to make further changes to the data model to accommodate national AND regional dex querying.