More database work
In order to code the connections correctly in Python, I went back to the database to get everything I need in order. My solution for the lack of National Dex queryability is to create a National Dex AND a Regional Dex for each game. Turns out, it seems to be a pretty good solution. Technically, there are two Pokedexes in each of the games that have the National Dex, so I'm not too mad about double storing. It is a bit of overhead I would have liked to avoid though. As I'm reshaping the data to fit the new schema, I'm finding some slight issues in the data validity. Artifacts of trying to store Pokedex data by generation, rather than by game. For example, Ultra Necrozma is in Sun and Moon at the moment. I think I'll make a quick GUI solution after this to navigate and remove Pokemon that don't appear in the respective Pokedex. I am now done moving Pokedex data for generation 1 through USUM. I'm going to finish this up before doing any more UI work, and I might also add Pokemon types and abilities as well before UI.