Renaming and UI organization
Renamed a bunch of controls to be a bit more generic. For example, as this is a Pokemon-focused application, I don't think every control needs to have 'pkmn' in it. The code should be quite a bit more readable now. There is still a bit more to be done in this area however. Also organized the UI controls into frames, to make future development a bit more manageable.