Updated DB function syntax and optimized import statements
Updated the UI to use the new StatsFrame class, and moved the cur_pokemon_id to the DB class. This may be a bit strange logically, but it shortens some of the function calls from the UI. My end goal with these changes is to make the UI code as slim and object-oriented as possible, so this brings us one step closer.