Added fail safes
Cleaned up  a few lines of code and added a couple of fail safes for game/dex selection. Now, if there is a dex for a game, it will automatically select the first dex in the list. Same with games. There is a bug now with the order in which the code is declared. This should be fixed as I break out the code into classes.