A bit of cleanup and documentation
Added a few comments to make the code more orderly. Removed the old Pokeball placeholder image, as the currently used placeholder is already in the database. Added a quick blurb about the project's purpose in the readme file. Organized and renamed a couple of functions for better self-documentation. Added a scrollbar to the Pokemon treeview for easier list navigation.