Search bar addition
Added a search bar to the Pokemon tree view. This allows users to quickly look up Pokemon, either by name or pokedex number. I haven't properly formatted it yet, but functionally it seems fine. Found a bug in the stats. Not sure if this is a data issue or a querying issue yet, but it appears that stat changes over the generations aren't properly displaying in the app.