Updated database through Gen 8
All Pokedexes from USUM -> Legends: Arceus have been added. It still only shows Pokemon stats, and that they are in the game, but this should be the hardest part. I'll need to do some QA/QC work after everything is in to be sure it is all valid, but this is certainly a good start.