Merged Frames into ViewerTab.py
The code was a bit cluttered and redundant, so I merged the code from the PokemonFrame and StatsFrame files into a new file, ViewerTab. Also, added a Notebook viewer to the UI. I think the final UI will look better if the viewer/editor can be accessed via a tabulated Notebook object. This way, each tab can be its own UI file, to keep things clean and organized.