Moved dex controls and added types interface
Dex controls have been moved to the selector subframe of the ViewerTab class. Additionally, Pokemon type icons have been added to the database, as well as the ERD. These icons are being pulled through the PokeDB class, and can be passed to the ViewerTab to display the appropriate Pokemon types under the main icon.