Enhanced stats UI and changed general UI style
Set the UI style to clam, to make everything look a little less plain. Added a self.max_stats variable to the ViewerTab to adjust the stat bars based on which game is being viewed. The color of the stat bars is also now being adjusted based on the values of the selected Pokemon's stats. Also fixed a few various bugs in the scripts.