UI Logic Cleanup
Opened the project on OSX and noticed several issues with the UI. This may be a result of questionable UI programming, so I have organized the code a bit better to remove some artifacts. Currently, the Pokemon TreeView object is displaying much too low on the UI, and I can't quite figure out why. Probably an issue with the grid logic, but I'll come back to this.