Rearranging UI and Shiny toggle
UI looked off, so it has been rearranged again. I grouped all of the selection variables together, minus the shiny toggle, which is newly added. I think it should stay separate from the rest of the selectors, as it is more organic to click the shiny toggle while looking at the actual Pokemon icon, rather than a bunch of tree views. Currently, the UI is a bit cluttered, which should be able to be fixed via padding here and there.