Finished organizing UI and added abilities
Arranged everything in the UI as I want it for now. It's a "good enough for me" kind of stop. Added abilities to the UI as well. Not sure if I'm going to work on the editor tab now, or maybe nix that idea altogether, but the view side of the application is officially done, minus some code cleanup.