Project cleanup
Zipped the Unused Icons folder to reduce the size/clutter.  Changed the database name to more accurately represent the rest of the project. Added screenshots to the new Screenshots folder to get a quick look at the interface without running anything. Added various type hints and comments to functions. Renamed various variables and functions for better clarity. Cleaned up the ERD a bit and moved it to the project root for better visibility. Fully revamped the README.md file to actually describe and define the project at hand. Removed a few functions from the EditorTab.py file which should no longer be needed, even when I eventually add editing functionality.